chore: consolidate shared test helpers into conftest.py (closes #44)
Moves _find_ffmpeg, _find_ffprobe, requires_ffmpeg, make_silent_mp3, and
make_jpeg into conftest.py so they are available to all test files without
duplication. Removes identical copies from test_audio.py, test_epub.py,
and test_integration.py.

Closes #44
